[fix] log sollte sich jetzt überall schließen
