+++ src/main/resources/github/weichware10/toolbox/gui/util/exit.png
